microservices/job_offers_scrapper/pracuj_subscriber/requirements.txt
requests
google-cloud-storage
uvicorn
sqlalchemy
pg8000
fastapi